Verschieben der Codierungsfunktion in die super Klasse
