+++ src/main/java/org/dhbw/Controller.java
@@ -6,6 +6,12 @@ import backend.database.DbCallerUser;
 import backend.database.DbConnector;
 import backend.usability.User;
 
+import javax.crypto.SecretKeyFactory;
+import javax.crypto.spec.PBEKeySpec;
+import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
+import java.security.spec.InvalidKeySpecException;
+import java.security.spec.KeySpec;
 import java.sql.SQLException;
 
 public class Controller {
@@ -31,5 +37,17 @@ public class Controller {
     // This method close a PopUp
     public void closeScene(){
 
+    }
+
+    public String encryptPassword(String password) throws InvalidKeySpecException, NoSuchAlgorithmException {
+        SecureRandom random = new SecureRandom();
+        byte[] salt = new byte[16];
+
+        random.nextBytes(salt);
+        KeySpec spec = new PBEKeySpec(password.toCharArray(), salt, 65536, 128);
+
+        SecretKeyFactory factory = SecretKeyFactory.getInstance("PBKDF2WithHmacSHA1");
+        String hash = String.valueOf(factory.generateSecret(spec).getEncoded());
+        return hash;
     }
 }
